Connected express backend
I connected the backend using the [instructions in the Create React App documentation](https://create-react-app.dev/docs/proxying-api-requests-in-development).
I created a new subdirectory `/backend` to hold the backend code.
The code in this subdirectory was adapted from the code that is
generated from `npx express-generator backend --no-view --git`.
This subdirectory is a new node.js project with a new `package.json`.
Therefore, to start the backend and frontend of the website, you have to
run `npm start` in both the top level directory and in the `/backend` directory.